C++ code: double -> float; Python: rewritten; runs, but erratic output. initializations?

+++ transs.cpp
@@ -71,13 +71,13 @@ int main(int argc, char* argv[]) {
       std::tie(coords, n_rows, n_cols) = Tools::IO::read_coords<double>(fname_input, 'C', Tools::range<std::size_t>(0, pc_max, 1));
     }
     // compute sigmas (and bandwidths) for every dimension
-    std::vector<double> sigmas(n_cols);
-    std::vector<double> bandwidths(n_cols);
+    std::vector<float> sigmas(n_cols);
+    std::vector<float> bandwidths(n_cols);
     std::vector<double> col_min(n_cols,  std::numeric_limits<double>::infinity());
     std::vector<double> col_max(n_cols, -std::numeric_limits<double>::infinity());
     {
       using namespace boost::accumulators;
-      using VarAcc = accumulator_set<double, features<tag::variance(lazy)>>;
+      using VarAcc = accumulator_set<float, features<tag::variance(lazy)>>;
       for (std::size_t j=0; j < n_cols; ++j) {
         VarAcc acc;
         for (std::size_t i=0; i < n_rows; ++i) {
@@ -99,12 +99,12 @@ int main(int argc, char* argv[]) {
       }
     }
     // compute transfer entropies
-    std::vector<std::vector<double>> T(n_cols, std::vector<double>(n_cols, 0.0));
+    std::vector<std::vector<float>> T(n_cols, std::vector<float>(n_cols, 0.0));
     {
       std::size_t y, x, n, i;
-      double p_s_y, p_s_x;
-      double s_xxy, s_xy, s_xx, s_x;
-      double tmp_xn, tmp_xn_tau, tmp_yn;
+      float p_s_y, p_s_x;
+      float s_xxy, s_xy, s_xx, s_x;
+      float tmp_xn, tmp_xn_tau, tmp_yn;
       #pragma omp parallel for default(none)\
                                private(y,x,n,i,p_s_y,p_s_x,s_xxy,s_xy,s_xx,s_x,tmp_xn,tmp_xn_tau,tmp_yn)\
                                firstprivate(n_cols,n_rows,tau)\
